fix(ui): rename back-navigation-text source files to fix typo
The component folder and export were already named correctly, but the
source files used back-nativation-text.

Fixes #2348
