+++ ui/apps/everest/src/components/back-navigation-text/back-navigation-text.tsx
@@ -1,6 +1,6 @@
 import { Box, IconButton, Stack, Typography } from '@mui/material';
 import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
-import { BackNavigationTextProps } from './back-nativation-text.types';
+import { BackNavigationTextProps } from './back-navigation-text.types';
 
 const BackNavigationText = ({
   text,
