+++ ui/apps/everest/src/components/back-navigation-text/index.ts
@@ -1,1 +1,1 @@
-export { default } from './back-nativation-text';
+export { default } from './back-navigation-text';
